Updated DocumentLayout.json and refactored product-related classes
This commit includes significant changes to the `DocumentLayout.json` file, including updating the `SelectedChildIndex` from 0 to 6, adding and removing several bookmarks in the `Children` array, updating the `ViewState` of the `README.md` document, and removing the `EditorCaption` property from several documents. In addition, the `Product` and `ProductService` classes in `Product.cs` and `ProductService.cs` respectively were removed. A new class `CategoryProductCountStatistic` was added to `CategoryProductCountStatistic.cs`. Lastly, the `using` statement in `TestComponent.razor.cs` was updated to use the `SIS_Technology_InterviewProject.Data.Product` namespace and the `CategoryProductCountStatistic` class was removed from `TestComponent.razor.cs`.